+++ src/components/GOLBoard/GOLBoard.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 
 import GOLRow from '../GOLRow'
+import GameOfLifeContext from '../../contexts/gol'
 
 import { Board } from '@sondregj/conway'
 
@@ -8,14 +9,32 @@ import styles from './GOLBoard.module.scss'
 
 interface GOLBoardProps {
     board: Board
+    setBoard: (board: Board) => void
 }
 
-const GOLBoard: React.FC<GOLBoardProps> = ({ board }) => {
+const GOLBoard: React.FC<GOLBoardProps> = ({ board, setBoard }) => {
+    const toggleCell = (x, y) => {
+        const newBoard: Board = {
+            width: board.width,
+            height: board.height,
+            cells: board.cells.map((row, yIndex) =>
+                row.map((cell, xIndex) =>
+                    xIndex === x && yIndex === y
+                        ? { ...cell, alive: !cell.alive }
+                        : { ...cell },
+                ),
+            ),
+        }
+        setBoard(newBoard)
+    }
+
     return (
         <div className={styles.board}>
-            {board.cells.map((row, y) => (
-                <GOLRow key={y} row={row} y={y} />
-            ))}
+            <GameOfLifeContext.Provider value={{ toggleCell }}>
+                {board.cells.map((row, y) => (
+                    <GOLRow key={y} row={row} y={y} />
+                ))}
+            </GameOfLifeContext.Provider>
         </div>
     )
 }
